Huge refactoring: all server-side unctions moved to XRouterServer class
