+++ src/xrouter/xrouterdef.h
@@ -9,6 +9,8 @@
 #include <queue>
 #include <memory>
 
+#define MIN_BLOCK 200
+
 //******************************************************************************
 //******************************************************************************
 namespace xrouter
